+++ bin/flg.dart
@@ -5,9 +5,10 @@ import 'package:args/command_runner.dart';
 import 'package:flg/src/commands/generate_command.dart';
 import 'package:flg/src/commands/init_command.dart';
 import 'package:flg/src/commands/setup_command.dart';
+import 'package:flg/src/commands/task_command.dart';
 import 'package:flg/src/utils/console_utils.dart';
 
-const String version = '1.0.0';
+const String version = '1.1.0';
 
 Future<void> main(List<String> arguments) async {
   final runner = CommandRunner<int>(
@@ -17,7 +18,8 @@ Future<void> main(List<String> arguments) async {
   )
     ..addCommand(InitCommand())
     ..addCommand(GenerateCommand())
-    ..addCommand(SetupCommand());
+    ..addCommand(SetupCommand())
+    ..addCommand(TaskCommand());
 
   // Add global flags
   runner.argParser